backend/requirements.txt
Pillow==9.5.0
easyocr==1.7.0
loguru==0.7.0
python-multipart==0.0.6
imagehash==4.3.1
fastapi==0.96.0
uvicorn[standard]==0.22.0
numpy==1.23.1

